requirements.txt
# Core Web3 and Blockchain
web3==6.11.3
eth-account==0.10.0
eth-typing==3.5.2
eth-utils==2.3.1

# Pydantic for data validation (with binary wheels)
pydantic==2.5.0
pydantic-core==2.14.5

# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1

# Flask for resource server
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

# Cryptography
cryptography==41.0.7
ecdsa==0.18.0

# Environment and configuration
python-dotenv==1.0.0

# HTTP requests
requests==2.31.0
httpx==0.25.2

# Streamlit for frontend (minimal dependencies)
streamlit==1.29.0
altair==5.2.0
pandas==2.1.4
numpy==1.26.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
